Are now using Recoil for statemangement. Must style button properly
