+++ package.json
@@ -13,6 +13,7 @@
     "axios": "^1.3.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "recoil": "^0.7.7",
     "sass": "^1.62.1"
   },
   "devDependencies": {
