+++ src/App.scss
@@ -5,20 +5,30 @@
 }
 
 
-.header{
+nav {
     width: 100vw;
     background: #192524;
     display: flex;
     align-items: center;
-    justify-content: center;
-    height: 7.5vh;
-    h1{
+    height: 10vh;
+    padding: 30px 25px;
+
+    h1 {
         color: #ace176;
+        margin-right: auto;
+
     }
+
+    .login {
+
+        margin: 0;
+    }
+
+
 }
 
-.content{
-    .login{
+.content {
+    .login {
         justify: right;
     }
 }
