--- src/Header.jsx
@@ -1,9 +1,0 @@
-export default function Header(){
-    
-    
-    return(
-        <div className="header">
-            <h1>SpotiMap</h1>
-        </div>
-    )
-}
