+++ src/Navbar.jsx
@@ -1,0 +1,12 @@
+import Login from "./Login";
+
+export default function Navbar() {
+
+
+    return (
+        <nav>
+            <h1>SpotiMap</h1>
+            <Login />
+        </nav>
+    )
+}
