+++ src/topArtistsState.js
@@ -1,0 +1,6 @@
+import { atom } from 'recoil'
+
+export const topArtistsState = atom({
+    key: 'topArtistsState',
+    default: [],
+})
